syscall: modernize the syscall layer with user-space stubs, public headers, pselect, and sigaltstack
Refactor the kernel syscall dispatch and add user-space stubs for
dirent, dup, epoll, ioctl, mmap, mqueue, poll, sched, shm, signal,
sleep, and time syscalls. Add pselect with atomic sigmask update
and sigaltstack for alternate signal stack configuration.

Add public headers under include/ and wire user-space syscall entry
points for both RISC-V ABIs.

+++ user/syscall/sleep.c
@@ -4,12 +4,7 @@
  * usleep() and sleep()
  */
 
-#include <stdio.h>
-#include <unistd.h>
-#include <stdarg.h>
-#include <errno.h>
 #include <pthread.h>
-
 #include <syscall.h>
 
 int usleep(useconds_t us)
@@ -21,5 +16,5 @@ int usleep(useconds_t us)
 unsigned int sleep(unsigned int seconds)
 {
 	pthread_testcancel();
-	return syscall1(SYSCALL_MSLEEP, (unsigned long)seconds * 1000UL);
+	return syscall1(SYSCALL_MSLEEP, seconds * 1000UL);
 }
